Language diversity data from Nettle (1998).

A long-form dataset containing language diversity observations from Nettle (1998). The data attempt to explain global patterns of language diversity and are most useful in wide format.

Nettle, D. (1998). Explaining Global Patterns of Language Diversity. Journal of Anthropological Archaeology, 17, 354–374.

language_diversity

Format

A data frame with 444 rows and 4 variables:

Continent

continent lables

Country

country labels

Measurement

global pattern measures

Value

values of measurements

Source

https://github.com/jvcasillas/untidydata